+++ src/components/bms/BMS1.java
@@ -1,4 +1,4 @@
-package bms;
+package components.bms;
 
 /**
  * {@code BMS} represented as a two-dimensional array.